Update schematicType selection and library names
The selected schematicType from the schematicTypeComboBox is now used to dispatch an action for selecting the schematic type, instead of relying on optionTypes keys. This will provide flexibility in terms of schematic type selection. Furthermore, the NodeModuleSearchUtil is now configured to use the Angular CLI "@angular/cli" instead of "@ngxs/cli", matching the intended target library. Ensured parameters are in kebab-case for compatibility.